requirements.txt
mavsdk==2.*
opencv-python 
pillow 
imagehash 
numpy 
pandas 
scipy 
matplotlib
pyyaml
